requirements.txt
google-genai
PyYAML
pydantic
rich
tqdm
pytest
black
mypy 